+++ CHANGELOG.md
@@ -5,7 +5,9 @@ Please add your entries according to this format.
 
 ## Unreleased
 
+- Add new regex property to exclude sourceSets from analysis (#328)
 - Fix custom KtFmt tasks not compatible with configuration cache (#290)
+- Fix custom source directories not correctly recognized (#292)
 
 ## Version 0.19.0 _(2024-07-03)_